+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `ServiceIntegration` `flink_external_postgresql` type
 - Remove `CA_CERT` secret key for `Grafana`, `OpenSearch`, `Redis`, and `Clickhouse`. Can't be used with these service types
 - Fix `ConnectionPool` doesn't check service user precondition
+- Move immutable fields validation from webhooks to CRD validation rules
 
 ## v0.19.0 - 2024-04-18